move records to separate files
