+++ airbyte-integrations/bases/base-typing-deduping/src/main/java/io/airbyte/integrations/base/destination/typing_deduping/ParsedCatalog.java
@@ -1,0 +1,18 @@
+package io.airbyte.integrations.base.destination.typing_deduping;
+
+import java.util.List;
+
+public record ParsedCatalog(List<StreamConfig> streams) {
+
+  public StreamConfig getStream(String namespace, String name) {
+    return streams.stream()
+        .filter(s -> s.id().originalNamespace().equals(namespace) && s.id().originalName().equals(name))
+        .findFirst()
+        .orElseThrow(() -> new IllegalArgumentException(String.format(
+            "Could not find stream %s.%s out of streams %s",
+            namespace,
+            name,
+            streams.stream().map(stream -> stream.id().originalNamespace() + "." + stream.id().originalName()).toList())));
+  }
+
+}
