+++ airbyte-integrations/bases/base-typing-deduping/src/main/java/io/airbyte/integrations/base/destination/typing_deduping/StreamConfig.java
@@ -1,0 +1,16 @@
+package io.airbyte.integrations.base.destination.typing_deduping;
+
+import io.airbyte.protocol.models.v0.DestinationSyncMode;
+import io.airbyte.protocol.models.v0.SyncMode;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Optional;
+
+public record StreamConfig(StreamId id,
+                           SyncMode syncMode,
+                           DestinationSyncMode destinationSyncMode,
+                           List<ColumnId> primaryKey,
+                           Optional<ColumnId> cursor,
+                           LinkedHashMap<ColumnId, AirbyteType> columns) {
+
+}
