+++ airbyte-integrations/connectors/destination-bigquery/src/test/java/io/airbyte/integrations/destination/bigquery/BigQueryRecordConsumerTest.java
@@ -10,8 +10,7 @@ import com.google.cloud.bigquery.BigQuery;
 import io.airbyte.commons.json.Jsons;
 import io.airbyte.integrations.base.DestinationConfig;
 import io.airbyte.integrations.base.FailureTrackingAirbyteMessageConsumer;
-import io.airbyte.integrations.base.TypingAndDedupingFlag;
-import io.airbyte.integrations.base.destination.typing_deduping.CatalogParser;
+import io.airbyte.integrations.base.destination.typing_deduping.ParsedCatalog;
 import io.airbyte.integrations.destination.bigquery.typing_deduping.BigQueryDestinationHandler;
 import io.airbyte.integrations.destination.bigquery.typing_deduping.BigQuerySqlGenerator;
 import io.airbyte.integrations.destination.bigquery.uploader.AbstractBigQueryUploader;
@@ -47,7 +46,7 @@ public class BigQueryRecordConsumerTest extends PerStreamStateMessageTest {
         "test-dataset-id",
         mock(BigQuerySqlGenerator.class),
         mock(BigQueryDestinationHandler.class),
-        new CatalogParser.ParsedCatalog(Collections.emptyList()));
+        new ParsedCatalog(Collections.emptyList()));
   }
 
   @Override
